requirements.txt
argparse
climetlab>=0.5.6
gridpp==0.6.0.dev16
matplotlib
netCDF4<1.6.0
numpy
pyyaml
scipy
tensorflow-gpu
tqdm
wheel>=0.22
PyYAML
